Rename folder to be in config instead of sessions
