+++ firebase-config/test-app/src/androidTest/kotlin/com/google/firebase/testing/config/FirebaseConfigTest.kt
